refactor: move error handling features to another file
Create ErrorHandler object
Move BridgeSizeError object to ErrorHandler
Change name: BridgeSizeError -> BridgeSize
