+++ src/ErrorHandler.js
@@ -1,0 +1,21 @@
+const { printInputErrorMessage } = require("./OutputView");
+const { makeBridge } = require("./BridgeMaker");
+const { generate } = require("./BridgeRandomNumberGenerator");
+const { ERROR } = require("./constants");
+
+const ErrorHandler = {
+  BridgeSize: {
+    true(app, _, readAgain) {
+      printInputErrorMessage(ERROR.WRONG_BRIDGE_SIZE);
+      readAgain(app);
+    },
+
+    false(app, bridgeSize, _) {
+      const bridge = makeBridge(bridgeSize, generate);
+
+      app.createGame(bridge);
+    },
+  },
+};
+
+module.exports = ErrorHandler;
