+++ src/InputHandler.js
@@ -1,27 +1,11 @@
 const { checkBridgeSizeInput } = require("./ErrorCase");
-const { printInputErrorMessage } = require("./OutputView");
-const { makeBridge } = require("./BridgeMaker");
-const { generate } = require("./BridgeRandomNumberGenerator");
-const { ERROR } = require("./constants");
-
-const BridgeSizeError = {
-  true(app, _, readAgain) {
-    printInputErrorMessage(ERROR.WRONG_BRIDGE_SIZE);
-    readAgain(app);
-  },
-
-  false(app, bridgeSize, _) {
-    const bridge = makeBridge(bridgeSize, generate);
-
-    app.createGame(bridge);
-  },
-};
+const { BridgeSize } = require("./ErrorHandler");
 
 const InputHandler = {
   handleBridgeSizeInput(app, input, readAgain) {
     const isError = checkBridgeSizeInput(input);
 
-    BridgeSizeError[isError](app, input, readAgain);
+    BridgeSize[isError](app, input, readAgain);
   },
 };
